+++ gegelatilib/include/log/LALogger.h
@@ -17,14 +17,17 @@ protected:
     /// keeps a given time to be able to show durations from that moment
     std::shared_ptr<std::chrono::time_point<std::chrono::system_clock, std::chrono::nanoseconds>> checkpoint;
 
-    /// returns the duration from a given beginning
+    /// \brief returns the duration from a given time to now
+    /// \param[in] begin time from which we want to compute duration
+    /// \return the duration from begin to now
     double getDurationFrom(std::chrono::time_point<std::chrono::system_clock, std::chrono::nanoseconds> &begin);
 
-    /// returns the current time value
+    /// \brief gets the current time value, for exemple to set checkpoint
+    /// \return the current time value
     std::chrono::time_point<std::chrono::system_clock, std::chrono::nanoseconds> getTime();
 
 public:
-    /// basic constructor setting cout as output, start and checkpoint as now
+    /// \brief basic constructor setting cout as output, start and checkpoint as now
     LALogger()
             : Logger(),
               start(std::make_shared<std::chrono::time_point<std::chrono::system_clock, std::chrono::nanoseconds>>(
@@ -32,24 +35,29 @@ public:
         chronoFromNow();
     };
 
-    /// constructor defining a given output and setting start and checkpoint as now
-    LALogger(std::ostream &stream) : Logger(stream),
+    /// \brief constructor defining a given output and setting start and checkpoint as now
+    /// \param[in] out the output stream we want to log things to
+    LALogger(std::ostream &out) : Logger(out),
                                      start(std::make_shared<std::chrono::time_point<std::chrono::system_clock, std::chrono::nanoseconds>>(
                                              getTime())) {
         chronoFromNow();
     };
 
-    /// updates checkpoint to now
+    /// \brief updates checkpoint to now
     void chronoFromNow();
 
-    /// called by the Learning Agent right after PopulateTPG is done
+    /// \brief method called by the Learning Agent right after PopulateTPG is done
+    /// \param[in] generationNumber the number of the current generation
+    /// \param[in] tpg the tpg we currently have
     virtual void logAfterPopulateTPG(uint64_t &generationNumber, TPG::TPGGraph &tpg) = 0;
 
-    /// called by the Learning Agent right after the evaluation is done
+    /// \brief method called by the Learning Agent right after the evaluation is done
+    /// \param[in] results scores of the evaluation
     virtual void
     logAfterEvaluate(std::multimap<std::shared_ptr<Learn::EvaluationResult>, const TPG::TPGVertex *> &results) = 0;
 
-    /// called by the Learning Agent right after the decimation is done
+    /// \brief method called by the Learning Agent right after the decimation is done
+    /// \param[in] tpg the tpg we currently have
     virtual void logAfterDecimate(TPG::TPGGraph &tpg) = 0;
 };